docgen/requirements.txt
langchain-groq
tree-sitter>=0.23.0
tree-sitter-python
chromadb
networkx
matplotlib
